requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-jwt==1.19.2
idna==3.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
ujson==5.4.0
urllib3==1.26.12
